cellGrob {patchwork}	R Documentation

Get a grob describing the content of an assemble_cell object

Description

Methods for this generic should be defined for all assemble_cell subclasses and should return a compliant gtable object ready to be combined with regular plot objects. In general it is best to call cell_table() on the object and add grobs to this as cell_table() will return a compliant gtable

Usage

cellGrob(x)

Arguments

x
An assemble_cell object

Value

A gtable object

[Package patchwork version 0.0.1 Index]