patchGrob {patchwork}	R Documentation

Get a grob describing the content of a patch object

Description

Methods for this generic should be defined for all patch subclasses and should return a compliant gtable object ready to be combined with regular plot objects. In general it is best to call patch_table() on the object and add grobs to this as patch_table() will return a compliant gtable

Usage

patchGrob(x, guides = "auto")

Arguments

x
An patch object

Value

A gtable object

[Package patchwork version 1.1.3 Index]