Convert a patchwork to a gtable

This function is the patchwork analogue of ggplot2::ggplotGrob() in that it takes an unevaluated patchwork object and fixate it into a gtable object to further manipulate directly.

patchworkGrob(x)

Arguments

x

A patchwork object

Value

A gtable object