Wrap arbitrary graphics in a patchwork-compliant patch

In order to add non-ggplot2 element to a patchwork they can be converted to a compliant representation using the wrap_elements() function. This allows you to position either grobs, ggplot objects, or patchwork objects in either the full area, the full plotting area (anything between and including the axis label), or the panel area (only the actual area where data is drawn). Further you can still add title, subtitle, tag, and caption using the same approach as with normal ggplots (using ggtitle() and labs()) as well as styling using theme(). For the latter, only the theme elements targeting plot margins and background as well as title, subtitle, etc styling will have an effect. If a patchwork or ggplot object is wrapped, it will be fixated in its state and will no longer respond to addition of styling, geoms, etc..

wrap_elements(panel = NULL, plot = NULL, full = NULL, clip = TRUE,
  ignore_tag = FALSE)

Arguments

panel, plot, full
A grob, ggplot, or patchwork object to add to the respective area.

clip
Should the grobs be clipped if expanding outside its area

ignore_tag
Should tags be ignored for this patch. This is relevant when using automatic tagging of plots and the content of the patch does not qualify for a tag.

Value

A wrapped_patch object