patchworkGrob {patchwork}	R Documentation

Convert a patchwork to a gtable

Description

This function is the patchwork analogue of ggplot2::ggplotGrob() in that it takes an unevaluated patchwork object and fixate it into a gtable object to further manipulate directly.

Usage

patchworkGrob(x)

Arguments

x
A patchwork object

Value

A gtable object

[Package patchwork version 1.0.0 Index]